Update jesus.mcfunction
Some changes to make the jesus anti-hack more reliable. Before the changes the anti-hack would think that a player is using jesus just from swimming and not resetting it afterwards. Therefore I made it so that the coah.t.jesus counter will count downwards whenever the player is inside of water to avoid false detections. It would also falsly detect jesus hacks when a player was on top of a boat, shifting out on carpets, slabs, trapdoors, lily pads etc... Therefore I made the block detection more precise. I also changed the detection time to be 6 seconds (120 ticks) instead of 240 ticks. This should be enough because of the changes I did regarding coah.t.jesus