Create nofall.mcfunction
A nofall anti-hack. When a player is using the nofall hack the "OnGround" tag stays at 1 even when they are in air. Therefore I'm checking if a player actually is in the air and then checking if it still says that they're on the ground. If they are in the air but OnGround at the same time they are using nofall hacks. I'm using a timer of 100 ticks (5 seconds) to make sure that the detection doesn't spam. Since I'm using this method we can even detect if a player who is using NoFall is just jumping a 1 block fall (on ground floor).